server/requirements.txt
flask
flask-cors
openai
flask-socketio
python-dotenv
langgraph
langchain-openai